+++ src/importFixer.ts
@@ -1,12 +1,38 @@
-import parseImport from 'parse-import-es6';
+import parseImport, { ImportDeclaration } from 'parse-import-es6';
 import * as vscode from 'vscode';
 import strip from 'parse-comment-es6';
 import { ImportObj } from './scanner';
-import { isIndexFile, isWin } from './help';
+import { isIndexFile, isWin, getImportOption } from './help';
+import ImportStatement, { EditChange } from './importStatement';
 const path = require('path');
 
+function getImportDeclaration(importedDefaultBinding, nameSpaceImport, namedImports,
+    importPath: string, position: vscode.Position): ImportDeclaration {
+    return {
+        importedDefaultBinding,
+        namedImports,
+        nameSpaceImport,
+        loc: {
+            start: {
+                line: position.line,
+                column: position.character,
+            },
+            end: {
+                line: position.line,
+                column: position.character,
+            },
+        },
+        range: null,
+        raw: '',
+        middleComments: [],
+        leadComments: [],
+        trailingComments: [],
+        moduleSpecifier: importPath,
+        error: 0,
+    }
+}
+
 export default class ImportFixer {
-    queto: string;
     eol: string;
     doc: vscode.TextDocument;
     importObj: ImportObj;
@@ -18,11 +44,6 @@ export default class ImportFixer {
         this.range = range;
         if (doc != null) {
             this.eol = doc.eol === vscode.EndOfLine.LF ? '\n' : '\r\n';
-        }
-        let queto = vscode.workspace.getConfiguration('js-import').get<string>('quote');
-        this.queto = `'`;
-        if (queto === 'doublequote') {
-            this.queto = `"`;
         }
     }
 
@@ -105,64 +126,61 @@ export default class ImportFixer {
         // TODO: here we can normalize moduleSpecifier
         const filteredImports = imports.filter(imp => imp.error === 0 && imp.moduleSpecifier === importPath);
 
+        let importStatement: ImportStatement = null;
         if (filteredImports.length === 0) {
-            let importStatement;
+            const position = this.getNewImportPositoin(imports);
             if (this.importObj.module.default) {
-                importStatement = this.getSingleLineImport(this.importObj.module.name, null, [], importPath, true);
+                importStatement = new ImportStatement(
+                    getImportDeclaration(this.importObj.module.name, null, [], importPath, position),
+                    getImportOption(this.eol),
+                );
             } else {
-                importStatement = this.getSingleLineImport(null, null, [this.importObj.module.name], importPath, true);
+                importStatement = new ImportStatement(
+                    getImportDeclaration(null, null, [this.importObj.module.name], importPath, position),
+                    getImportOption(this.eol),
+                );
             }
-            this.insertNewImport(imports, importStatement);
         } else {
-            // TODO: merge imports
-            this.replaceOldImport(filteredImports[0], importPath);
-        }
-    }
-
-    public replaceOldImport(imp, importPath) {
-        let importStatement;
-        if (this.importObj.module.default) {
-            if (imp.importedDefaultBinding !== null && imp.importedDefaultBinding === this.importObj.module.name) {
-                // TODO: we can format code
-                return;
-            } else if (imp.importedDefaultBinding !== null && imp.importedDefaultBinding !== this.importObj.module.name) {
-                // error , two default import
-                return;
+            // TODO: merge import
+            const imp = filteredImports[0];
+            if (this.importObj.module.default) {
+                if (imp.importedDefaultBinding !== null && imp.importedDefaultBinding === this.importObj.module.name) {
+                    // TODO: we can format code
+                    return;
+                } else if (imp.importedDefaultBinding !== null && imp.importedDefaultBinding !== this.importObj.module.name) {
+                    // error , two default import
+                    return;
+                } else {
+                    // imp.importedDefaultBinding === null
+                    importStatement = new ImportStatement(
+                        Object.assign(imp, { importedDefaultBinding: this.importObj.module.name }),
+                        getImportOption(this.eol),
+                    );
+                }
             } else {
-                // imp.importedDefaultBinding === null
-                if (imp.loc.start.line < imp.loc.end.line && imp.namedImports.length !== 0) {
-                    this.getMultipleLineImport(imp, this.importObj.module.name, imp.nameSpaceImport, imp.namedImports, importPath);
+                if (imp.nameSpaceImport !== null) {
+                    // error
                     return;
                 }
-                // TODO: if imp.middleComments is not empty, wo should extract all middle comment into end to avoid missing comments
-                importStatement = this.getSingleLineImport(this.importObj.module.name, imp.nameSpaceImport, imp.namedImports, importPath, false);
+                if (imp.namedImports.includes(this.importObj.module.name)) {
+                    // TODO: we can format code
+                    return;
+                }
+                importStatement = new ImportStatement(
+                    Object.assign(imp, { namedImports: imp.namedImports.concat([this.importObj.module.name]) }),
+                    getImportOption(this.eol),
+                );
             }
-        } else {
-            if (imp.nameSpaceImport !== null) {
-                // error
-                return;
-            }
-            if (imp.namedImports.includes(this.importObj.module.name)) {
-                // TODO: we can format code
-                return;
-            }
-            if (imp.loc.start.line < imp.loc.end.line) {
-                this.getMultipleLineImport(imp, imp.importedDefaultBinding, imp.nameSpaceImport,
-                    imp.namedImports.concat([this.importObj.module.name]), importPath);
-                return;
-            }
-            importStatement = this.getSingleLineImport(imp.importedDefaultBinding, imp.nameSpaceImport,
-                imp.namedImports.concat([this.importObj.module.name]), importPath, false);
         }
+
+        const iec: EditChange = importStatement.getEditChange();
         let edit: vscode.WorkspaceEdit = new vscode.WorkspaceEdit();
-        edit.replace(this.doc.uri, new vscode.Range(imp.loc.start.line, imp.loc.start.column,
-            imp.loc.end.line, imp.loc.end.column), importStatement);
+        edit.replace(this.doc.uri, new vscode.Range(iec.startLine, iec.startColumn, iec.endLine, iec.endColumn), iec.text);
         this.deleteWordRange(edit);
         vscode.workspace.applyEdit(edit);
     }
 
-    public insertNewImport(imports, importStatement) {
-        let newText = importStatement;
+    public getNewImportPositoin(imports) {
         let position: vscode.Position = null;
         let pos = vscode.workspace.getConfiguration('js-import').get<string>('insertPosition') || 'last';
         if (pos !== 'first' && pos !== 'last') {
@@ -173,7 +191,6 @@ export default class ImportFixer {
             if (imp.trailingComments.length === 0) {
                 position = new vscode.Position(imp.loc.end.line + 1, 0);
             } else {
-                newText = this.eol + newText;
                 position = new vscode.Position(imp.trailingComments[imp.trailingComments.length - 1].loc.end.line + 1, 0);
             }
         }
@@ -182,7 +199,6 @@ export default class ImportFixer {
             const comments = strip(this.doc.getText(), { comment: true, range: true, loc: true, raw: true }).comments;
 
             if (comments.length === 0) {
-                newText = newText + this.eol;
                 position = new vscode.Position(0, 0);
             } else {
                 // exculde the first leading comment of the first import, if exist 'flow' 'Copyright' 'LICENSE'
@@ -196,121 +212,18 @@ export default class ImportFixer {
                     }
                 }
                 if (index === comments.length) {
-                    // should insert after every comment
-                    newText = this.eol + newText;
                     position = new vscode.Position(comment.loc.end.line + 1, 0);
                 } else {
                     if (index === 0) {
-                        newText = newText + this.eol;
                         position = new vscode.Position(0, 0);
                     } else {
-                        // should insert after previous current comment
-                        newText = this.eol + newText;
                         comment = comments[index - 1];
                         position = new vscode.Position(comment.loc.end.line + 1, 0);
                     }
                 }
             }
         }
-        let edit: vscode.WorkspaceEdit = new vscode.WorkspaceEdit();
-        edit.insert(this.doc.uri, position, newText);
-        this.deleteWordRange(edit);
-        vscode.workspace.applyEdit(edit);
-    }
-
-    /**
-     * https://tc39.github.io/ecma262/#prod-ImportClause
-     * @param imp
-     * @param importPath
-     * @param endline
-     */
-    public getSingleLineImport(importedDefaultBinding, nameSpaceImport, namedImports, importPath: string, endline = false) {
-        let commaDangleImport = vscode.workspace.getConfiguration('js-import').get<string>('commaDangleImport') || 'never';
-        let namedImportsText = null;
-        if (commaDangleImport === 'always' && namedImports.length !== 0) {
-            namedImportsText = `${namedImports.join(', ')},`
-        } else {
-            namedImportsText = `${namedImports.join(', ')}`
-        }
-
-        if (importedDefaultBinding !== null && nameSpaceImport !== null) {
-            return `import ${importedDefaultBinding}, { ${nameSpaceImport} } from ${this.queto}${importPath}${this.queto}${endline ? this.eol : ''};`
-        } else if (importedDefaultBinding !== null && namedImports.length !== 0) {
-            return `import ${importedDefaultBinding}, { ${namedImportsText} } from ${this.queto}${importPath}${this.queto}${endline ? this.eol : ''};`
-        } else if (importedDefaultBinding !== null) {
-            return `import ${importedDefaultBinding} from ${this.queto}${importPath}${this.queto};${endline ? this.eol : ''}`
-        } else if (nameSpaceImport !== null) {
-            return `import ${nameSpaceImport} from ${this.queto}${importPath}${this.queto};${endline ? this.eol : ''}`
-        } else if (namedImports.length !== 0) {
-            return `import { ${namedImportsText} } from ${this.queto}${importPath}${this.queto};${endline ? this.eol : ''}`
-        } else {
-            // do nothing
-        }
-    }
-
-    public getMultipleLineImport(imp, importedDefaultBinding, nameSpaceImport, namedImports, importPath) {
-        // TODO: split multiple lines if exceed character per line (use a parameter setting)
-        let newText = 'import ';
-        let startline = imp.loc.start.line;
-        let startcolumn = imp.loc.start.column;
-        let endline = imp.loc.end.line;
-        let endcolumn = imp.loc.end.column;
-        if (importedDefaultBinding != null) {
-            newText += importedDefaultBinding + ', {';
-        } else {
-            newText += ' {'
-        }
-
-        // handle comment of import or default import
-        const defaultImportComments = imp.middleComments.filter(comment => comment.identifier.identifier === importedDefaultBinding || comment.identifier.type === 'Import');
-        if (defaultImportComments.length != 0) {
-            newText += ' ';
-        }
-        defaultImportComments.forEach(element => {
-            newText += element.raw;
-            startcolumn = Math.min(startcolumn, element.loc.start.column);
-        });
-
-        namedImports.forEach((element, index) => {
-            newText += this.eol;
-            // handle comment before identifier in previous line
-            const beforeNamedImportsComments = imp.middleComments.filter(comment => comment.identifier.identifier === element && comment.loc.start.line < comment.identifier.loc.start.line);
-            beforeNamedImportsComments.forEach(comment => {
-                newText += `    ${comment.raw}${this.eol}`;
-            });
-            let commaDangleImport = vscode.workspace.getConfiguration('js-import').get<string>('commaDangleImport') || 'never';
-            if (commaDangleImport === 'always' || commaDangleImport === 'always-multiline') {
-                newText += `    ${element},`;
-            } else {
-                if (index === namedImports.length - 1) {
-                    newText += `    ${element}`;
-                } else {
-                    newText += `    ${element},`;
-                }
-            }
-            // handle comment after identifier
-            const afterNamedImportsComments = imp.middleComments.filter(comment => comment.identifier.identifier === element && comment.loc.start.line >= comment.identifier.loc.start.line);
-            if (afterNamedImportsComments.length != 0) {
-                newText += ' ';
-            }
-            afterNamedImportsComments.forEach(comment => {
-                newText += comment.raw;
-            });
-        });
-
-        newText += `${this.eol}} from ${this.queto}${importPath}${this.queto};`;
-        const comments = imp.middleComments.filter(comment => comment.identifier.type === 'From' || comment.identifier.type === 'ModuleSpecifier');
-        if (comments.length != 0) {
-            newText += ' ';
-        }
-        comments.forEach(comment => {
-            newText += comment.raw;
-            endcolumn = Math.max(endcolumn, comment.loc.end.column);
-        });
-        let edit: vscode.WorkspaceEdit = new vscode.WorkspaceEdit();
-        edit.replace(this.doc.uri, new vscode.Range(startline, startcolumn, endline, endcolumn), newText);
-        this.deleteWordRange(edit);
-        vscode.workspace.applyEdit(edit);
+        return position;
     }
 
     public deleteWordRange(edit: vscode.WorkspaceEdit) {